Web Scraping
This Project is both a skill demonstrator and a problem solver.
In my college we run into a big problem in the day on which our final semester results gets published from the server to the website, which is, overloading on the server.
which in turn, makes pulling data from the server very slow.
I provided a solution using selenium in python to automatically go to the website and do the following:
1- Submit the student ID
2- Choosing the right link for the required results (Because it shows two links, one for the first half of the year and the second is the full year)
3- Scraping final results (Student's name, Results, Department)
4- analyzing performance based on collected Data in powerBI ( in the results folder )