Subway Analysis
This project is one of my favourites, It's very Simple, yet, very informative.

In this project I am solving a critical problem here in egypt, which is: Determining the average time taken to get from point a to point b in cairo subway.

why is this a problem anyway?
because it's randomness!

you read it right, In Egypt this is a common problem because you can never say that a trip from station a to station b is a solid 2 minutes, that'd be a wrong sentence to make

because you did not cosider the randomness that happens in transitional stations (stations in which people travel from a line to another line)
nor you considered the randomness that is due to just being random ( train being late, people holding the door for their friends ....etc)

in this simple statistical projects all you see in the python file is raw data I collected myself in 4 days in order to make this project valid.

Skills Used:
	1.python
	2.Numpy
	3.Hypothesis Testing
	4.Bootstraping

you can find an article about this whole project in here: https://theveryamateuredatascientist.blogspot.com/2022/07/estimating-required-te-to-move-from.html